Modifications to Mediator theme
I pulled the mediator theme, and changed some of it's bits around. First, I removed the author feature because it would have become a hindrance when there's multiple authors. Next I made an optional featured post feature. I changed the site settings to point to the proper urls, and changed the footer of the posts.

The website is using the Jekyll feature of Github pages. http://jekyllrb.com/

Posts are written in markdown offering flexibility for those who don't care to learn web shit.